# PERF: Merge empty frame