@jreback 2022-02-09
lgtm will merge on green.

tip to avoid conflicts in the release notes; add a note *not* at the bottom, add in middle somewhere